feat : 사칙 연산 계산 기능 숫자, 연산자 부분 구현

+++ src/main/java/jumble_jump/domain/token/NumberToken.java
@@ -3,5 +3,5 @@ package jumble_jump.domain.token;
 import jumble_jump.util.Token;
 
 public interface NumberToken extends Token {
-    int getNumber();
+    double getNumber();
 }
